app/requirements.txt
django==3.1.7
django-autocomplete-light==3.8.2
django-countries==7.0
django-crispy-forms==1.11.1
django-debug-toolbar==3.2
django-environ==0.4.5
djangorestframework==3.12.2
django-extensions==3.1.1
django-object-actions==3.0.1
django-tempus-dominus==5.1.2.13
django-user-agents==0.4.0
feedparser==6.0.2
ipython==7.21.0
gunicorn==19.9.0
requests==2.25.1